perf(server): bound the per-connection WebSocket outbound backlog
The per-connection outbound channel was unbounded: a client that stopped
reading made the server buffer every broadcast notification forever,
growing memory without limit and silently defeating the broadcast-lag
accounting that is supposed to surface slow consumers.

Bound the channel at 256 frames and switch all producers to try_send.
When the queue is full the notifier signals the main loop through a watch
channel, which closes the connection — the same policy the heartbeat
already applies to stalled peers. Responses, pings, and the shutdown
Close frame follow the same path.

Also moves the WebSocket test module to websocket/tests.rs so the
implementation file stays reviewable, and adds an end-to-end regression
test that floods a non-reading client until the server closes it.

Closes #1996
